Ex. 12.18 and 12.19 done. Configure the Nginx server to be in front of todo-backend.
12.18: Add the service todo-backend to the docker-compose file `todo-app/docker-compose.dev.yml` in development mode.

Add a new location to the `nginx.dev.conf` so that requests to `/api` are proxied to the backend.

12.19: Connect the services, todo-frontend with todo-backend.

The development environment is now fully functional, that is:

a. All features of the todo app work;
b. One can edit the source files and the changes take effect by reloading the app (the hot reloading may or may not work...);
c. Frontend should access the backend through Nginx, so the requests should be done to http://localhost:8080/api/todos